Refactor base toolbox code de-coupled from app & config into tool_util.
This was originally decoupled by me for this purpose and this code was in galaxy-lib years ago, but it just never made it into galaxy-tool-util. This commit clears up that oversight I think.
