+++ lib/galaxy/tool_util/toolbox/filters/__init__.py
@@ -22,7 +22,7 @@ class FilterFactory:
         self.default_filters = dict(tool=[_not_hidden, _handle_authorization], section=[], label=[])
         # Add dynamic filters to these default filters.
         config = toolbox.app.config
-        self.__base_modules = listify(getattr(config, "toolbox_filter_base_modules", "galaxy.tools.filters"))
+        self.__base_modules = listify(getattr(config, "toolbox_filter_base_modules", "galaxy.tool_util.filters"))
         self.__init_filters("tool", getattr(config, "tool_filters", ""), self.default_filters)
         self.__init_filters("section", getattr(config, "tool_section_filters", ""), self.default_filters)
         self.__init_filters("label", getattr(config, "tool_label_filters", ""), self.default_filters)
