+++ lib/galaxy/tool_util/toolbox/filters/examples.py.sample
