+++ lib/galaxy/tool_util/toolbox/integrated_panel.py
