+++ lib/galaxy/tool_util/toolbox/panel.py
