+++ lib/galaxy/tools/filters/__init__.py
@@ -1,5 +1,5 @@
 """Deprecated module for toolbox filters.
 
 Filters placed in this module will still work - but filters should be
-moved to lib/galaxy/tools/toolbox/filters.
+moved to lib/galaxy/tool_util/toolbox/filters.
 """
