+++ packages/app/requirements.txt
@@ -14,7 +14,6 @@ gxformat2
 Mako
 sqlitedict
 svgwrite
-sortedcontainers
 Markdown
 refgenconf>=0.7.0,<0.10.0 # https://github.com/galaxyproject/galaxy/issues/11601
 Fabric3
