+++ packages/app/setup.py
@@ -61,9 +61,6 @@ PACKAGES = [
     'galaxy.tools.imp_exp',
     'galaxy.tools.parameters',
     'galaxy.tools.search',
-    'galaxy.tools.toolbox',
-    'galaxy.tools.toolbox.filters',
-    'galaxy.tools.toolbox.lineages',
     'galaxy.tools.util',
     'galaxy.tools.util.galaxyops',
     'galaxy.tool_shed',
